products/seo-agent/requirements.txt
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-frontmatter==1.1.0
toml==0.10.2
PyYAML==6.0.2
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.3
